pyproject.toml
[project]
name = "python-langchain-test"
version = "0.1.0"
description = "AI Research Assistant Agent with LangChain and LangSmith"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.0",
    "tavily-python>=0.5.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
